+++ Datadog/E2ETests/NTP/KronosE2ETests.swift
@@ -10,7 +10,6 @@ class KronosE2ETests: E2ETests {
     /// The logger sending logs on Kronos execution. These logs are available in Mobile Integrations org.
     private var logger: Logger! // swiftlint:disable:this implicitly_unwrapped_optional
     /// The logger sending telemetry on internal Kronos execution. These logs are available in Mobile Integrations org.
-    private var telemetryLogger: Logger! // swiftlint:disable:this implicitly_unwrapped_optional
     private let queue = DispatchQueue(label: "kronos-monitor-queue")
 
     override func setUp() {
@@ -19,31 +18,11 @@ class KronosE2ETests: E2ETests {
             .builder
             .set(loggerName: "kronos-e2e")
             .build()
-        telemetryLogger = Logger.builder
-            .set(loggerName: "kronos-e2e-internal-telemetry")
-            .sendNetworkInfo(true)
-            .build()
     }
 
     override func tearDown() {
         logger = nil
-        telemetryLogger = nil
         super.tearDown()
-    }
-
-    /// Creates kronos monitor for checking connections to all IPs resolved from NTP pool and sending additional telemetry on their statuses.
-    private func createKronosMonitor() -> KronosMonitor? {
-        if #available(iOS 14.2, *) {
-            let monitor = KronosInternalMonitor(
-                queue: queue,
-                connectionMonitor: IPConnectionMonitor(queue: queue)
-            )
-            // Here we redirect IM's logger to E2E Kronos logger (`telemetryLogger`) to send data to Mobile Integrations org, not IM's org
-            monitor.export(to: InternalMonitor(sdkLogger: telemetryLogger))
-            return monitor
-        } else {
-            return nil
-        }
     }
 
     ///  TODO: RUMM-1859: Add E2E tests for monitoring Kronos in nightly tests
@@ -81,7 +60,6 @@ class KronosE2ETests: E2ETests {
             KronosClock.sync(
                 from: pool,
                 samples: numberOfSamplesForEachIP,
-                monitor: createKronosMonitor(),
                 first: { date, offset in // this closure could not be called if all samples to all servers resulted with failure
                     result.firstReceivedDate = date
                     result.firstReceivedOffset = offset
@@ -129,96 +107,6 @@ class KronosE2ETests: E2ETests {
                     "serverOffset_lastReceived": result.lastReceivedOffset,
                 ])
             }
-        }
-    }
-
-    ///  TODO: RUMM-1859: Add E2E tests for monitoring Kronos in nightly tests
-    func test_kronos_ntp_client_queries_both_ipv4_and_ipv6_ips() { // E2E:wip
-        /// The result of `KronosNTPClient.query(pool:)`.
-        struct KronosNTPClientQueryResult {
-            /// Partial offsets received for each NTP packet sent to each resolved IP.
-            var receivedOffsets: [TimeInterval?] = []
-            /// Expected number of NTP packets to send.
-            var expectedNumberOfSamples = 0
-            /// Actual number of NTP packets that completed.
-            var numberOfCompletedSamples = 0
-        }
-
-        func performKronosNTPClientQuery() -> KronosNTPClientQueryResult {
-            let testTimeout: TimeInterval = 30
-            let monitor = createKronosMonitor()
-
-            // Given
-            let pool = "2.datadog.pool.ntp.org" // a pool resolved to multiple IPv4 and IPv6 addresses (e.g. 4 + 4)
-            let numberOfSamplesForEachIP = 2 // exchange only 2 samples with each resolved IP - to run test quick
-
-            // Each IP (each server) is asked in parallel, but samples are obtained sequentially.
-            // Here we compute individual sample timeout, to ensure that all (parallel) servers complete querying their (sequential) samples
-            // below `testTimeout` with assuming -30% margin. This should guarantee no flakiness on test timeout.
-            let timeoutForEachSample = (testTimeout / Double(numberOfSamplesForEachIP)) * 0.7
-
-            // When
-            let completionExpectation = expectation(description: "It completes all samples for all IPs")
-            var result = KronosNTPClientQueryResult()
-
-            monitor?.notifySyncStart(from: pool) // must be notified by hand because normally it's called from `KronosClock.sync()`
-
-            KronosNTPClient()
-                .query(
-                    pool: pool,
-                    numberOfSamples: numberOfSamplesForEachIP,
-                    maximumServers: .max, // query all resolved IPs in the pool - to include both IPv4 and IPv6
-                    timeout: timeoutForEachSample,
-                    monitor: monitor
-                ) { offset, completed, total in
-                    result.receivedOffsets.append(offset)
-                    result.numberOfCompletedSamples = completed
-                    result.expectedNumberOfSamples = total
-
-                    if completed == total {
-                        monitor?.notifySyncEnd(serverOffset: offset) // must be notified by hand because normally it's called from `KronosClock.sync()`
-                        completionExpectation.fulfill()
-                    }
-                }
-
-            // Then
-
-            // We don't expect receiving timeout on `completionExpectation`. Number of samples and individual sample timeout
-            // is configured in a way that lets `KronosNTPClient` always fulfill the `completionExpectation`.
-            // In worst case, it can fulfill it, with recording only `nil` offsets, which will mean receiving timeouts
-            // or error on all NTP queries.
-            waitForExpectations(timeout: testTimeout)
-
-            return result
-        }
-
-        // Run test:
-        let result = measure(resourceName: DD.PerfSpanName.fromCurrentMethodName()) {
-            performKronosNTPClientQuery()
-        }
-
-        // Report result:
-        if result.receivedOffsets.contains(where: { offset in offset != nil }) {
-            // We consider `KronosNTPClient.query(pool:)` result to be consistent if it received at least one offset.
-            let receivedOffsets: [String] = result.receivedOffsets.map { offset in
-                if let offset = offset {
-                    return "\(offset)"
-                } else {
-                    return "(nil)"
-                }
-            }
-            logger.info(
-                "KronosNTPClient.query(pool:) completed with consistent result receiving \(result.numberOfCompletedSamples)/\(result.expectedNumberOfSamples) NTP packets",
-                attributes: [
-                    "offsets_received": receivedOffsets
-                ]
-            )
-        } else {
-            // Inconsistent result may correspond to flaky execution, e.g. if network was unreachable or if **all** NTP calls received timeout.
-            // We track inconsistent result as WARN log that will be watched by E2E monitor.
-            logger.warn(
-                "KronosNTPClient.query(pool:) completed with inconsistent result receiving \(result.numberOfCompletedSamples)/\(result.expectedNumberOfSamples) NTP packets"
-            )
         }
     }
 }