+++ Sources/Datadog/RUM/Instrumentation/Resources/URLSessionRUMResourcesHandler.swift
@@ -42,7 +42,6 @@ internal class URLSessionRUMResourcesHandler: URLSessionInterceptionHandler, RUM
                 url: url,
                 httpMethod: RUMMethod(httpMethod: interception.request.httpMethod),
                 kind: RUMResourceType(request: interception.request),
-                isFirstPartyRequest: interception.isFirstPartyRequest,
                 spanContext: interception.spanContext.map {
                     .init(
                         traceID: String($0.traceID.rawValue),