+++ Sources/Datadog/RUMMonitor.swift
@@ -349,7 +349,6 @@ public class RUMMonitor: DDRUMMonitor, RUMCommandSubscriber {
                 url: request.url?.absoluteString ?? "unknown_url",
                 httpMethod: RUMMethod(httpMethod: request.httpMethod),
                 kind: RUMResourceType(request: request),
-                isFirstPartyRequest: nil,
                 spanContext: nil
             )
         )
@@ -368,7 +367,6 @@ public class RUMMonitor: DDRUMMonitor, RUMCommandSubscriber {
                 url: url.absoluteString,
                 httpMethod: .get,
                 kind: nil,
-                isFirstPartyRequest: nil,
                 spanContext: nil
             )
         )
@@ -388,7 +386,6 @@ public class RUMMonitor: DDRUMMonitor, RUMCommandSubscriber {
                 url: urlString,
                 httpMethod: httpMethod,
                 kind: nil,
-                isFirstPartyRequest: nil,
                 spanContext: nil
             )
         )