+++ Tests/DatadogTests/Datadog/Core/Upload/DataUploadWorkerTests.swift
@@ -227,14 +227,12 @@ class DataUploadWorkerTests: XCTestCase {
 
     // MARK: - Notifying Upload Progress
 
-    func testWhenDataIsBeingUploaded_itPrintsUploadProgressInformationAndSendsErrorsThroughInternalMonitoring() {
+    func testWhenDataIsBeingUploaded_itPrintsUploadProgressInformation() {
         let previousUserLogger = userLogger
         defer { userLogger = previousUserLogger }
 
         let mockUserLoggerOutput = LogOutputMock()
         userLogger = .mockWith(logOutput: mockUserLoggerOutput)
-
-        let mockSDKLoggerOutput = LogOutputMock()
 
         // Given
         writer.write(value: ["key": "value"])
@@ -253,10 +251,7 @@ class DataUploadWorkerTests: XCTestCase {
             dataUploader: mockDataUploader,
             uploadConditions: .alwaysUpload(),
             delay: DataUploadDelay(performance: UploadPerformanceMock.veryQuickInitialUpload),
-            featureName: randomFeatureName,
-            internalMonitor: InternalMonitor(
-                sdkLogger: .mockWith(logOutput: mockSDKLoggerOutput)
-            )
+            featureName: randomFeatureName
         )
 
         wait(for: [startUploadExpectation], timeout: 0.5)
@@ -264,7 +259,7 @@ class DataUploadWorkerTests: XCTestCase {
 
         // Then
         let expectedSummary = randomUploadStatus.needsRetry ? "not delivered, will be retransmitted" : "accepted, won't be retransmitted"
-        XCTAssertEqual(mockUserLoggerOutput.allRecordedLogs.count, 3)
+        XCTAssertEqual(mockUserLoggerOutput.allRecordedLogs.count, 2)
 
         XCTAssertEqual(
             mockUserLoggerOutput.allRecordedLogs[0].message,
@@ -277,17 +272,114 @@ class DataUploadWorkerTests: XCTestCase {
             "   → (\(randomFeatureName)) \(expectedSummary): \(randomUploadStatus.userDebugDescription)",
             "Batch completion information should be printed to `userLogger`. All captured logs:\n\(mockUserLoggerOutput.dumpAllRecordedLogs())"
         )
-        XCTAssertEqual(
-            mockUserLoggerOutput.allRecordedLogs[2].message,
-            randomUploadStatus.userErrorMessage,
-            "An error should be printed to `userLogger`. All captured logs:\n\(mockUserLoggerOutput.dumpAllRecordedLogs())"
+    }
+
+    func testWhenDataIsBeingUploaded_itPrintsUnauthoriseMessage_toUserLogger() {
+        let previousUserLogger = userLogger
+        defer { userLogger = previousUserLogger }
+
+        let mockUserLoggerOutput = LogOutputMock()
+        userLogger = .mockWith(logOutput: mockUserLoggerOutput)
+
+        // Given
+        writer.write(value: ["key": "value"])
+
+        let randomUploadStatus: DataUploadStatus = .mockWith(error: .unauthorized)
+
+        // When
+        let startUploadExpectation = self.expectation(description: "Upload has started")
+        var mockDataUploader = DataUploaderMock(uploadStatus: randomUploadStatus)
+        mockDataUploader.onUpload = { startUploadExpectation.fulfill() }
+
+        let worker = DataUploadWorker(
+            queue: uploaderQueue,
+            fileReader: reader,
+            dataUploader: mockDataUploader,
+            uploadConditions: .alwaysUpload(),
+            delay: DataUploadDelay(performance: UploadPerformanceMock.veryQuickInitialUpload),
+            featureName: .mockRandom()
         )
 
-        XCTAssertEqual(mockSDKLoggerOutput.allRecordedLogs.count, 1)
+        wait(for: [startUploadExpectation], timeout: 0.5)
+        worker.cancelSynchronously()
+
+        // Then
+        XCTAssertEqual(mockUserLoggerOutput.allRecordedLogs.count, 3)
+
         XCTAssertEqual(
-            mockSDKLoggerOutput.allRecordedLogs[0].message,
-            randomUploadStatus.internalMonitoringError?.message,
-            "An error should be send to `sdkLogger` for internal monitoring. All captured logs:\n\(mockSDKLoggerOutput.dumpAllRecordedLogs())"
+            mockUserLoggerOutput.allRecordedLogs[2].message,
+            "⚠️ Make sure that the provided token still exists and you're targeting the relevant Datadog site.",
+            "An error should be printed to `userLogger`. All captured logs:\n\(mockUserLoggerOutput.dumpAllRecordedLogs())"
+        )
+    }
+
+    func testWhenDataIsBeingUploaded_itPrintsHTTPErrorMessage_toTelemetry() {
+        // Given
+        let mockTelemetry = TelemetryMock()
+
+        writer.write(value: ["key": "value"])
+        let randomUploadStatus: DataUploadStatus = .mockWith(error: .httpError(statusCode: 500))
+
+        // When
+        let startUploadExpectation = self.expectation(description: "Upload has started")
+        var mockDataUploader = DataUploaderMock(uploadStatus: randomUploadStatus)
+        mockDataUploader.onUpload = { startUploadExpectation.fulfill() }
+
+        let worker = DataUploadWorker(
+            queue: uploaderQueue,
+            fileReader: reader,
+            dataUploader: mockDataUploader,
+            uploadConditions: .alwaysUpload(),
+            delay: DataUploadDelay(performance: UploadPerformanceMock.veryQuickInitialUpload),
+            featureName: .mockRandom(),
+            telemetry: mockTelemetry
+        )
+
+        wait(for: [startUploadExpectation], timeout: 0.5)
+        worker.cancelSynchronously()
+
+        // Then
+        XCTAssertEqual(mockTelemetry.errors.count, 1)
+
+        XCTAssertEqual(
+            mockTelemetry.errors.first?.message,
+            "Data upload finished with status code: 500",
+            "An error should be send to internal telemetry. \(mockTelemetry)"
+        )
+    }
+
+    func testWhenDataIsBeingUploaded_itPrintsNetworkErrorMessage_toTelemetry() {
+        // Given
+        let mockTelemetry = TelemetryMock()
+
+        writer.write(value: ["key": "value"])
+        let randomUploadStatus: DataUploadStatus = .mockWith(error: .networkError(error: .mockAny()))
+
+        // When
+        let startUploadExpectation = self.expectation(description: "Upload has started")
+        var mockDataUploader = DataUploaderMock(uploadStatus: randomUploadStatus)
+        mockDataUploader.onUpload = { startUploadExpectation.fulfill() }
+
+        let worker = DataUploadWorker(
+            queue: uploaderQueue,
+            fileReader: reader,
+            dataUploader: mockDataUploader,
+            uploadConditions: .alwaysUpload(),
+            delay: DataUploadDelay(performance: UploadPerformanceMock.veryQuickInitialUpload),
+            featureName: .mockRandom(),
+            telemetry: mockTelemetry
+        )
+
+        wait(for: [startUploadExpectation], timeout: 0.5)
+        worker.cancelSynchronously()
+
+        // Then
+        XCTAssertEqual(mockTelemetry.errors.count, 1)
+
+        XCTAssertEqual(
+            mockTelemetry.errors.first?.message,
+            #"Data upload finished with error - Error Domain=abc Code=0 "(null)""#,
+            "An error should be send to internal telemetry. \(mockTelemetry)"
         )
     }