+++ Tests/DatadogTests/Datadog/Mocks/RUMFeatureMocks.swift
@@ -28,9 +28,15 @@ extension RUMFeature {
     static func mockWith(
         directories: FeatureDirectories,
         configuration: FeaturesConfiguration.RUM = .mockAny(),
-        dependencies: FeaturesCommonDependencies = .mockAny()
+        dependencies: FeaturesCommonDependencies = .mockAny(),
+        telemetry: Telemetry? = nil
     ) -> RUMFeature {
-        return RUMFeature(directories: directories, configuration: configuration, commonDependencies: dependencies)
+        return RUMFeature(
+            directories: directories,
+            configuration: configuration,
+            commonDependencies: dependencies,
+            telemetry: telemetry
+        )
     }
 
     /// Mocks the feature instance which performs uploads to mocked `DataUploadWorker`.
@@ -98,14 +104,16 @@ extension RUMTelemetry: AnyMockable {
         applicationID: String = .mockAny(),
         source: String = .mockAnySource(),
         dateProvider: DateProvider = SystemDateProvider(),
-        dateCorrector: DateCorrectorType = DateCorrectorMock()
+        dateCorrector: DateCorrectorType = DateCorrectorMock(),
+        sampler: Sampler = .init(samplingRate: 100)
     ) -> Self {
         .init(
             sdkVersion: sdkVersion,
             applicationID: applicationID,
             source: source,
             dateProvider: dateProvider,
-            dateCorrector: dateCorrector
+            dateCorrector: dateCorrector,
+            sampler: sampler
         )
     }
 }
@@ -354,7 +362,6 @@ extension RUMStartResourceCommand: AnyMockable, RandomMockable {
             url: url,
             httpMethod: httpMethod,
             kind: kind,
-            isFirstPartyRequest: isFirstPartyRequest,
             spanContext: spanContext
         )
     }
@@ -631,7 +638,7 @@ internal struct NoOpRUMViewUpdatesThrottler: RUMViewUpdatesThrottlerType {
     }
 }
 
-func mockNoOpSessionListerner() -> RUMSessionListener {
+func mockNoOpSessionListener() -> RUMSessionListener {
     return { _, _ in }
 }
 
@@ -656,6 +663,7 @@ extension RUMScopeDependencies {
         applicationVersion: String = .mockAny(),
         sdkVersion: String = .mockAny(),
         source: String = "ios",
+        firstPartyURLsFilter: FirstPartyURLsFilter = FirstPartyURLsFilter(hosts: []),
         eventBuilder: RUMEventBuilder = RUMEventBuilder(eventsMapper: .mockNoOp()),
         eventOutput: RUMEventOutput = RUMEventOutputMock(),
         rumUUIDGenerator: RUMUUIDGenerator = DefaultRUMUUIDGenerator(),
@@ -663,7 +671,7 @@ extension RUMScopeDependencies {
         crashContextIntegration: RUMWithCrashContextIntegration? = nil,
         ciTest: RUMCITest? = nil,
         viewUpdatesThrottlerFactory: @escaping () -> RUMViewUpdatesThrottlerType = { NoOpRUMViewUpdatesThrottler() },
-        onSessionStart: @escaping RUMSessionListener = mockNoOpSessionListerner()
+        onSessionStart: @escaping RUMSessionListener = mockNoOpSessionListener()
     ) -> RUMScopeDependencies {
         return RUMScopeDependencies(
             rumApplicationID: rumApplicationID,
@@ -678,6 +686,7 @@ extension RUMScopeDependencies {
             applicationVersion: applicationVersion,
             sdkVersion: sdkVersion,
             source: source,
+            firstPartyURLsFilter: firstPartyURLsFilter,
             eventBuilder: eventBuilder,
             eventOutput: eventOutput,
             rumUUIDGenerator: rumUUIDGenerator,
@@ -706,6 +715,7 @@ extension RUMScopeDependencies {
         applicationVersion: String? = nil,
         sdkVersion: String? = nil,
         source: String? = nil,
+        firstPartyUrls: Set<String>? = nil,
         eventBuilder: RUMEventBuilder? = nil,
         eventOutput: RUMEventOutput? = nil,
         rumUUIDGenerator: RUMUUIDGenerator? = nil,
@@ -728,6 +738,7 @@ extension RUMScopeDependencies {
             applicationVersion: applicationVersion ?? self.applicationVersion,
             sdkVersion: sdkVersion ?? self.sdkVersion,
             source: source ?? self.source,
+            firstPartyURLsFilter: firstPartyUrls.map { .init(hosts: $0) } ?? self.firstPartyURLsFilter,
             eventBuilder: eventBuilder ?? self.eventBuilder,
             eventOutput: eventOutput ?? self.eventOutput,
             rumUUIDGenerator: rumUUIDGenerator ?? self.rumUUIDGenerator,
@@ -860,7 +871,6 @@ extension RUMResourceScope {
             dateCorrection: dateCorrection,
             url: url,
             httpMethod: httpMethod,
-            isFirstPartyResource: isFirstPartyResource,
             resourceKindBasedOnRequest: resourceKindBasedOnRequest,
             spanContext: spanContext,
             onResourceEventSent: onResourceEventSent,