+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changes
 
+* [IMPROVEMENT] Allow manually tracked resources in RUM Sessions to detect first party hosts. 
 * [IMPROVEMENT] Add tracing sampling rate. See [#851][]
 
 # 1.11.0-beta2 / 05-04-2022